@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :library do
     name 'Some Library'
   end
