@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :term do
     dtstart '2015-05-27 15:01:15'
     dtend '2015-05-27 15:01:15'
