@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,88 +10,105 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150527220242) do
+ActiveRecord::Schema.define(version: 2018_03_19_235103) do
+
+  create_table "active_storage_attachments", force: :cascade do |t|
+    t.string "name", null: false
+    t.string "record_type", null: false
+    t.integer "record_id", null: false
+    t.integer "blob_id", null: false
+    t.datetime "created_at", null: false
+    t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
+    t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
+  end
+
+  create_table "active_storage_blobs", force: :cascade do |t|
+    t.string "key", null: false
+    t.string "filename", null: false
+    t.string "content_type"
+    t.text "metadata"
+    t.bigint "byte_size", null: false
+    t.string "checksum", null: false
+    t.datetime "created_at", null: false
+    t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+  end
 
   create_table "calendars", force: :cascade do |t|
     t.datetime "dtstart"
     t.datetime "dtend"
-    t.integer  "location_id"
-    t.string   "summary"
-    t.text     "description"
-    t.boolean  "closed"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
+    t.integer "location_id"
+    t.string "summary"
+    t.text "description"
+    t.boolean "closed"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["location_id"], name: "index_calendars_on_location_id"
   end
-
-  add_index "calendars", ["location_id"], name: "index_calendars_on_location_id"
 
   create_table "friendly_id_slugs", force: :cascade do |t|
-    t.string   "slug",                      null: false
-    t.integer  "sluggable_id",              null: false
-    t.string   "sluggable_type", limit: 50
-    t.string   "scope"
+    t.string "slug", null: false
+    t.integer "sluggable_id", null: false
+    t.string "sluggable_type", limit: 50
+    t.string "scope"
     t.datetime "created_at"
+    t.index ["slug", "sluggable_type", "scope"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope", unique: true
+    t.index ["slug", "sluggable_type"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type"
+    t.index ["sluggable_id"], name: "index_friendly_id_slugs_on_sluggable_id"
+    t.index ["sluggable_type"], name: "index_friendly_id_slugs_on_sluggable_type"
   end
 
-  add_index "friendly_id_slugs", ["slug", "sluggable_type", "scope"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope", unique: true
-  add_index "friendly_id_slugs", ["slug", "sluggable_type"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type"
-  add_index "friendly_id_slugs", ["sluggable_id"], name: "index_friendly_id_slugs_on_sluggable_id"
-  add_index "friendly_id_slugs", ["sluggable_type"], name: "index_friendly_id_slugs_on_sluggable_type"
-
   create_table "libraries", force: :cascade do |t|
-    t.string   "name"
-    t.string   "slug"
+    t.string "name"
+    t.string "slug"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
   create_table "locations", force: :cascade do |t|
-    t.string   "name"
-    t.string   "slug"
-    t.boolean  "keeps_hours", default: false
-    t.integer  "library_id"
-    t.datetime "created_at",                  null: false
-    t.datetime "updated_at",                  null: false
+    t.string "name"
+    t.string "slug"
+    t.boolean "keeps_hours", default: false
+    t.integer "library_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["library_id"], name: "index_locations_on_library_id"
   end
 
-  add_index "locations", ["library_id"], name: "index_locations_on_library_id"
-
   create_table "node_mappings", force: :cascade do |t|
-    t.integer  "node_id"
-    t.integer  "mapped_id"
-    t.string   "mapped_type"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
+    t.integer "node_id"
+    t.integer "mapped_id"
+    t.string "mapped_type"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "spreadsheets", force: :cascade do |t|
-    t.string   "attachment_id"
-    t.string   "attachment_filename"
-    t.string   "attachment_content_type"
-    t.integer  "attachment_size"
-    t.string   "status"
-    t.datetime "created_at",              null: false
-    t.datetime "updated_at",              null: false
+    t.string "attachment_id"
+    t.string "attachment_filename"
+    t.string "attachment_content_type"
+    t.integer "attachment_size"
+    t.string "status"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "term_hours", force: :cascade do |t|
-    t.integer  "term_id"
-    t.integer  "location_id"
-    t.text     "data"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
+    t.integer "term_id"
+    t.integer "location_id"
+    t.text "data"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["location_id"], name: "index_term_hours_on_location_id"
+    t.index ["term_id"], name: "index_term_hours_on_term_id"
   end
-
-  add_index "term_hours", ["location_id"], name: "index_term_hours_on_location_id"
-  add_index "term_hours", ["term_id"], name: "index_term_hours_on_term_id"
 
   create_table "terms", force: :cascade do |t|
     t.datetime "dtstart"
     t.datetime "dtend"
-    t.string   "name"
-    t.boolean  "holiday",    default: false
-    t.datetime "created_at",                 null: false
-    t.datetime "updated_at",                 null: false
+    t.string "name"
+    t.boolean "holiday", default: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
 end
