@@ -10,33 +10,33 @@ RSpec.describe LocationsController, type: :controller do
     let(:location) { create(:location) }
 
     it 'should allow access to #show' do
-      get :show, id: location, library_id: location.library
+      get :show, params: { id: location, library_id: location.library }
     end
 
     it 'should deny access to #new' do
-      expect { get :new, library_id: location.library }.to raise_error CanCan::AccessDenied
+      expect { get :new, params: { library_id: location.library } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #edit' do
-      expect { get :edit, id: location, library_id: location.library }.to raise_error CanCan::AccessDenied
+      expect { get :edit, params:  { id: location, library_id: location.library } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #update' do
-      expect { post :update, id: location, library_id: location.library }.to raise_error CanCan::AccessDenied
+      expect { post :update, params: { id: location, library_id: location.library } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #destroy' do
-      expect { delete :destroy, id: location, library_id: location.library }.to raise_error CanCan::AccessDenied
+      expect { delete :destroy, params: { id: location, library_id: location.library } }.to raise_error CanCan::AccessDenied
     end
 
     describe 'GET #hours_v1' do
       it 'assigns @when' do
-        get :hours_v1, format: :json, when: 'today', id: location, library_id: location.library
+        get :hours_v1, format: :json, params: { when: 'today', id: location, library_id: location.library }
         expect(assigns(:when)).to eq Time.zone.now.beginning_of_day
       end
 
       it 'assigns the hours for the location' do
-        get :hours_v1, format: :json, when: 'today', id: location, library_id: location.library
+        get :hours_v1, format: :json, params: { when: 'today', id: location, library_id: location.library }
         expect(assigns(:hours).first.first).to be_a MissingCalendar
         expect(assigns(:hours).first.first.dtstart).to eq Time.zone.now.beginning_of_day
       end
@@ -44,13 +44,13 @@ RSpec.describe LocationsController, type: :controller do
 
     describe 'GET #hours' do
       it 'assigns @range from explicit dates' do
-        get :hours, format: :json, from: Time.zone.now.beginning_of_week.to_s, to: Time.zone.now.end_of_week.to_s, id: location, library_id: location.library
+        get :hours, format: :json, params: { from: Time.zone.now.beginning_of_week.to_s, to: Time.zone.now.end_of_week.to_s, id: location, library_id: location.library }
         expect(assigns(:range).begin).to eq Time.zone.now.beginning_of_week.to_date
         expect(assigns(:range).end).to eq Time.zone.now.end_of_week.to_date
       end
 
       it 'assigns the hours for the location' do
-        get :hours, format: :json, id: location, library_id: location.library
+        get :hours, format: :json, params: { id: location, library_id: location.library }
         expect(assigns(:hours).first.first).to be_a MissingCalendar
         expect(assigns(:hours).first.first.dtstart).to eq Time.zone.now.beginning_of_day
       end
@@ -84,14 +84,14 @@ RSpec.describe LocationsController, type: :controller do
   describe 'GET #show' do
     it 'assigns the requested location as @location' do
       location = library.locations.create! valid_attributes
-      get :show, library_attributes.merge(id: location.to_param), valid_session
+      get :show, params: library_attributes.merge(id: location.to_param), session: valid_session
       expect(assigns(:location)).to eq(location)
     end
   end
 
   describe 'GET #new' do
     it 'assigns a new location as @location' do
-      get :new, library_attributes, valid_session
+      get :new, params: library_attributes, session: valid_session
       expect(assigns(:location)).to be_a_new(Location)
     end
   end
@@ -99,7 +99,7 @@ RSpec.describe LocationsController, type: :controller do
   describe 'GET #edit' do
     it 'assigns the requested location as @location' do
       location = library.locations.create! valid_attributes
-      get :edit, library_attributes.merge(id: location.to_param), valid_session
+      get :edit, params: library_attributes.merge(id: location.to_param), session: valid_session
       expect(assigns(:location)).to eq(location)
     end
   end
@@ -108,30 +108,30 @@ RSpec.describe LocationsController, type: :controller do
     context 'with valid params' do
       it 'creates a new Location' do
         expect do
-          post :create, library_attributes.merge(location: valid_attributes), valid_session
+          post :create, params: library_attributes.merge(location: valid_attributes), session: valid_session
         end.to change(Location, :count).by(1)
       end
 
       it 'assigns a newly created location as @location' do
-        post :create, library_attributes.merge(location: valid_attributes), valid_session
+        post :create, params: library_attributes.merge(location: valid_attributes), session: valid_session
         expect(assigns(:location)).to be_a(Location)
         expect(assigns(:location)).to be_persisted
       end
 
       it 'redirects to the created location' do
-        post :create, library_attributes.merge(location: valid_attributes), valid_session
+        post :create, params: library_attributes.merge(location: valid_attributes), session: valid_session
         expect(response).to redirect_to([library, Location.last])
       end
     end
 
     context 'with invalid params' do
       it 'assigns a newly created but unsaved location as @location' do
-        post :create, library_attributes.merge(location: invalid_attributes), valid_session
+        post :create, params: library_attributes.merge(location: invalid_attributes), session: valid_session
         expect(assigns(:location)).to be_a_new(Location)
       end
 
       it "re-renders the 'new' template" do
-        post :create, library_attributes.merge(location: invalid_attributes), valid_session
+        post :create, params: library_attributes.merge(location: invalid_attributes), session: valid_session
         expect(response).to render_template('new')
       end
     end
@@ -145,20 +145,20 @@ RSpec.describe LocationsController, type: :controller do
 
       it 'updates the requested location' do
         location = library.locations.create! valid_attributes
-        put :update, library_attributes.merge(id: location.to_param, location: new_attributes), valid_session
+        put :update, params: library_attributes.merge(id: location.to_param, location: new_attributes), session: valid_session
         location.reload
         expect(location.name).to eq 'New Location Name'
       end
 
       it 'assigns the requested location as @location' do
         location = library.locations.create! valid_attributes
-        put :update, library_attributes.merge(id: location.to_param, location: valid_attributes), valid_session
+        put :update, params: library_attributes.merge(id: location.to_param, location: valid_attributes), session: valid_session
         expect(assigns(:location)).to eq(location)
       end
 
       it 'redirects to the location' do
         location = library.locations.create! valid_attributes
-        put :update, library_attributes.merge(id: location.to_param, location: valid_attributes), valid_session
+        put :update, params: library_attributes.merge(id: location.to_param, location: valid_attributes), session: valid_session
         expect(response).to redirect_to([library, location])
       end
     end
@@ -166,13 +166,13 @@ RSpec.describe LocationsController, type: :controller do
     context 'with invalid params' do
       it 'assigns the location as @location' do
         location = library.locations.create! valid_attributes
-        put :update, library_attributes.merge(id: location.to_param, location: invalid_attributes), valid_session
+        put :update, params: library_attributes.merge(id: location.to_param, location: invalid_attributes), session: valid_session
         expect(assigns(:location)).to eq(location)
       end
 
       it "re-renders the 'edit' template" do
         location = library.locations.create! valid_attributes
-        put :update, library_attributes.merge(id: location.to_param, location: invalid_attributes), valid_session
+        put :update, params: library_attributes.merge(id: location.to_param, location: invalid_attributes), session: valid_session
         expect(response).to render_template('edit')
       end
     end
@@ -182,13 +182,13 @@ RSpec.describe LocationsController, type: :controller do
     it 'destroys the requested location' do
       location = library.locations.create! valid_attributes
       expect do
-        delete :destroy, library_attributes.merge(id: location.to_param), valid_session
+        delete :destroy, params: library_attributes.merge(id: location.to_param), session: valid_session
       end.to change(Location, :count).by(-1)
     end
 
     it 'redirects to the locations list' do
       location = library.locations.create! valid_attributes
-      delete :destroy, library_attributes.merge(id: location.to_param), valid_session
+      delete :destroy, params: library_attributes.merge(id: location.to_param), session: valid_session
       expect(response).to redirect_to(library_url(location.library))
     end
   end
