@@ -25,14 +25,14 @@ RSpec.describe SpreadsheetsController, type: :controller do
 
   describe 'with an anonymous user' do
     let(:user) { nil }
-    let(:spreadsheet) { Spreadsheet.create! attachment_id: '123' }
+    let(:spreadsheet) { Spreadsheet.create! attachment: StringIO.new('xyz') }
 
     it 'should deny access to #index' do
       expect { get :index }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #show' do
-      expect { get :show, id: spreadsheet }.to raise_error CanCan::AccessDenied
+      expect { get :show, params: { id: spreadsheet } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #new' do
@@ -40,15 +40,15 @@ RSpec.describe SpreadsheetsController, type: :controller do
     end
 
     it 'should deny access to #edit' do
-      expect { get :edit, id: spreadsheet }.to raise_error CanCan::AccessDenied
+      expect { get :edit, params: { id: spreadsheet } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #update' do
-      expect { post :update, id: spreadsheet }.to raise_error CanCan::AccessDenied
+      expect { post :update, params: { id: spreadsheet } }.to raise_error CanCan::AccessDenied
     end
 
     it 'should deny access to #destroy' do
-      expect { delete :destroy, id: spreadsheet }.to raise_error CanCan::AccessDenied
+      expect { delete :destroy, params: { id: spreadsheet } }.to raise_error CanCan::AccessDenied
     end
   end
 
@@ -84,7 +84,7 @@ RSpec.describe SpreadsheetsController, type: :controller do
   describe 'GET #index' do
     it 'assigns all spreadsheets as @spreadsheets' do
       spreadsheet = Spreadsheet.create! valid_attributes
-      get :index, {}, valid_session
+      get :index, params: {}, session: valid_session
       expect(assigns(:spreadsheets)).to eq([spreadsheet])
     end
   end
@@ -92,14 +92,14 @@ RSpec.describe SpreadsheetsController, type: :controller do
   describe 'GET #show' do
     it 'assigns the requested spreadsheet as @spreadsheet' do
       spreadsheet = Spreadsheet.create! valid_attributes
-      get :show, { id: spreadsheet.to_param }, valid_session
+      get :show, params: { id: spreadsheet.to_param }, session: valid_session
       expect(assigns(:spreadsheet)).to eq(spreadsheet)
     end
   end
 
   describe 'GET #new' do
     it 'assigns a new spreadsheet as @spreadsheet' do
-      get :new, {}, valid_session
+      get :new, params: {}, session: valid_session
       expect(assigns(:spreadsheet)).to be_a_new(Spreadsheet)
     end
   end
@@ -107,7 +107,7 @@ RSpec.describe SpreadsheetsController, type: :controller do
   describe 'GET #edit' do
     it 'assigns the requested spreadsheet as @spreadsheet' do
       spreadsheet = Spreadsheet.create! valid_attributes
-      get :edit, { id: spreadsheet.to_param }, valid_session
+      get :edit, params: { id: spreadsheet.to_param }, session: valid_session
       expect(assigns(:spreadsheet)).to eq(spreadsheet)
     end
   end
@@ -116,30 +116,30 @@ RSpec.describe SpreadsheetsController, type: :controller do
     context 'with valid params' do
       it 'creates a new Spreadsheet' do
         expect do
-          post :create, { spreadsheet: valid_attributes }, valid_session
+          post :create, params: { spreadsheet: valid_attributes }, session: valid_session
         end.to change(Spreadsheet, :count).by(1)
       end
 
       it 'assigns a newly created spreadsheet as @spreadsheet' do
-        post :create, { spreadsheet: valid_attributes }, valid_session
+        post :create, params: { spreadsheet: valid_attributes }, session: valid_session
         expect(assigns(:spreadsheet)).to be_a(Spreadsheet)
         expect(assigns(:spreadsheet)).to be_persisted
       end
 
       it 'redirects to the created spreadsheet' do
-        post :create, { spreadsheet: valid_attributes }, valid_session
+        post :create, params: { spreadsheet: valid_attributes }, session: valid_session
         expect(response).to redirect_to(Spreadsheet.last)
       end
     end
 
     context 'with invalid params' do
       it 'assigns a newly created but unsaved spreadsheet as @spreadsheet' do
-        post :create, { spreadsheet: invalid_attributes }, valid_session
+        post :create, params: { spreadsheet: invalid_attributes }, session: valid_session
         expect(assigns(:spreadsheet)).to be_a_new(Spreadsheet)
       end
 
       it "re-renders the 'new' template" do
-        post :create, { spreadsheet: invalid_attributes }, valid_session
+        post :create, params: { spreadsheet: invalid_attributes }, session: valid_session
         expect(response).to render_template('new')
       end
     end
@@ -149,13 +149,13 @@ RSpec.describe SpreadsheetsController, type: :controller do
     it 'destroys the requested spreadsheet' do
       spreadsheet = Spreadsheet.create! valid_attributes
       expect do
-        delete :destroy, { id: spreadsheet.to_param }, valid_session
+        delete :destroy, params: { id: spreadsheet.to_param }, session: valid_session
       end.to change(Spreadsheet, :count).by(-1)
     end
 
     it 'redirects to the spreadsheets list' do
       spreadsheet = Spreadsheet.create! valid_attributes
-      delete :destroy, { id: spreadsheet.to_param }, valid_session
+      delete :destroy, params: { id: spreadsheet.to_param }, session: valid_session
       expect(response).to redirect_to(spreadsheets_url)
     end
   end
